console: Refactor application collaborator edit view
Remove 'api' usage in the component
