+++ pkg/webui/console/views/application-collaborator-edit/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,179 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React from 'react'
-import { connect } from 'react-redux'
-import bind from 'autobind-decorator'
-import { Container, Col, Row } from 'react-grid-system'
-import { replace } from 'connected-react-router'
+import ApplicationCollaboratorEdit from './application-collaborator-edit'
+import connect from './connect'
 
-import api from '@console/api'
+const ConnectedCollaboratorEdit = connect(ApplicationCollaboratorEdit)
 
-import { withBreadcrumb } from '@ttn-lw/components/breadcrumbs/context'
-import PageTitle from '@ttn-lw/components/page-title'
-import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
-import toast from '@ttn-lw/components/toast'
-
-import withRequest from '@ttn-lw/lib/components/with-request'
-
-import CollaboratorForm from '@console/components/collaborator-form'
-
-import sharedMessages from '@ttn-lw/lib/shared-messages'
-import PropTypes from '@ttn-lw/lib/prop-types'
-
-import { getCollaborator } from '@console/store/actions/collaborators'
-
-import {
-  selectSelectedApplicationId,
-  selectApplicationRights,
-  selectApplicationPseudoRights,
-  selectApplicationRightsFetching,
-  selectApplicationRightsError,
-} from '@console/store/selectors/applications'
-import {
-  selectUserCollaborator,
-  selectOrganizationCollaborator,
-  selectCollaboratorFetching,
-  selectCollaboratorError,
-} from '@console/store/selectors/collaborators'
-
-const isUser = collaborator => collaborator.ids && 'user_ids' in collaborator.ids
-
-@connect(
-  function(state, props) {
-    const appId = selectSelectedApplicationId(state)
-
-    const { collaboratorId, collaboratorType } = props.match.params
-
-    const collaborator =
-      collaboratorType === 'user'
-        ? selectUserCollaborator(state)
-        : selectOrganizationCollaborator(state)
-    const fetching = selectApplicationRightsFetching(state) || selectCollaboratorFetching(state)
-    const error = selectApplicationRightsError(state) || selectCollaboratorError(state)
-
-    return {
-      collaboratorId,
-      collaboratorType,
-      collaborator,
-      appId,
-      rights: selectApplicationRights(state),
-      pseudoRights: selectApplicationPseudoRights(state),
-      fetching,
-      error,
-    }
-  },
-  (dispatch, ownProps) => ({
-    getCollaborator(appId, collaboratorId, isUser) {
-      dispatch(getCollaborator('application', appId, collaboratorId, isUser))
-    },
-    redirectToList(appId) {
-      dispatch(replace(`/applications/${appId}/collaborators`))
-    },
-  }),
-  (stateProps, dispatchProps, ownProps) => ({
-    ...stateProps,
-    ...dispatchProps,
-    ...ownProps,
-    getCollaborator: () =>
-      dispatchProps.getCollaborator(
-        stateProps.appId,
-        stateProps.collaboratorId,
-        stateProps.collaboratorType === 'user',
-      ),
-    redirectToList: () => dispatchProps.redirectToList(stateProps.appId),
-  }),
-)
-@withRequest(
-  ({ getCollaborator }) => getCollaborator(),
-  ({ fetching, collaborator }) => fetching || !Boolean(collaborator),
-)
-@withBreadcrumb('apps.single.collaborators.edit', function(props) {
-  const { appId, collaboratorId, collaboratorType } = props
-
-  return (
-    <Breadcrumb
-      path={`/applications/${appId}/collaborators/${collaboratorType}/${collaboratorId}`}
-      content={sharedMessages.edit}
-    />
-  )
-})
-export default class ApplicationCollaboratorEdit extends React.Component {
-  static propTypes = {
-    appId: PropTypes.string.isRequired,
-    collaborator: PropTypes.collaborator.isRequired,
-    collaboratorId: PropTypes.string.isRequired,
-    pseudoRights: PropTypes.rights.isRequired,
-    redirectToList: PropTypes.func.isRequired,
-    rights: PropTypes.rights.isRequired,
-  }
-
-  state = {
-    error: '',
-  }
-
-  @bind
-  async handleSubmit(updatedCollaborator) {
-    const { appId } = this.props
-
-    await api.application.collaborators.update(appId, updatedCollaborator)
-  }
-
-  handleSubmitSuccess() {
-    toast({
-      message: sharedMessages.collaboratorUpdateSuccess,
-      type: toast.types.SUCCESS,
-    })
-  }
-
-  @bind
-  async handleDelete() {
-    const { collaborator, collaboratorId, redirectToList, appId } = this.props
-    const collaborator_type = isUser(collaborator) ? 'user' : 'organization'
-
-    const collaborator_ids = {
-      [`${collaborator_type}_ids`]: {
-        [`${collaborator_type}_id`]: collaboratorId,
-      },
-    }
-    const updatedCollaborator = {
-      ids: collaborator_ids,
-    }
-
-    try {
-      await api.application.collaborators.remove(appId, updatedCollaborator)
-      toast({
-        message: sharedMessages.collaboratorDeleteSuccess,
-        type: toast.types.SUCCESS,
-      })
-      redirectToList(appId)
-    } catch (error) {
-      await this.setState({ error })
-    }
-  }
-
-  render() {
-    const { collaborator, collaboratorId, rights, pseudoRights, redirectToList } = this.props
-
-    return (
-      <Container>
-        <PageTitle title={sharedMessages.collaboratorEdit} values={{ collaboratorId }} />
-        <Row>
-          <Col lg={8} md={12}>
-            <CollaboratorForm
-              error={this.state.error}
-              onSubmit={this.handleSubmit}
-              onSubmitSuccess={this.handleSubmitSuccess}
-              onDelete={this.handleDelete}
-              onDeleteSuccess={redirectToList}
-              collaborator={collaborator}
-              pseudoRights={pseudoRights}
-              rights={rights}
-              update
-            />
-          </Col>
-        </Row>
-      </Container>
-    )
-  }
-}
+export { ConnectedCollaboratorEdit as default, ApplicationCollaboratorEdit }
